Move slice-configurator to slices/
