+++ src/slices/app/app-slice.ts
@@ -4,8 +4,8 @@ import { createSlice, PayloadAction, type WithSlice } from '@reduxjs/toolkit';
 import { PersistConfig } from 'redux-persist';
 import { getEnv } from '../../app/environment.js';
 import type { RootState, ThunkResult } from '../../app/store.js';
-import { buildSliceConfigurator, SliceConfigurator } from '../../middleware/slice-configurator.js';
 import { Team } from '../../models/team.js';
+import { buildSliceConfigurator, SliceConfigurator } from '../slice-configurator.js';
 import { addTeam } from '../team/team-slice.js';
 
 const env = getEnv();
