+++ src/slices/game/game-slice.ts
@@ -2,10 +2,10 @@
 
 import { createAsyncThunk, createSlice, PayloadAction, type WithSlice } from '@reduxjs/toolkit';
 import { RootState, ThunkPromise, ThunkResult } from '../../app/store.js';
-import { buildSliceConfigurator, SliceConfigurator } from '../../middleware/slice-configurator.js';
 import { Game, GameDetail, Games, GameStatus } from '../../models/game.js';
 import { selectCurrentUserId } from '../auth/auth-slice.js';
 import { gameCompleted, gameSetupCompleted, selectLiveGameById } from '../live/live-slice.js';
+import { buildSliceConfigurator, SliceConfigurator } from '../slice-configurator.js';
 import { GamePayload } from './game-action-types.js';
 import {
   loadGame,
