+++ src/slices/live/composed-reducer.ts
@@ -3,7 +3,7 @@
 import { createNextState } from '@reduxjs/toolkit';
 import { Reducer } from 'redux';
 import { PersistConfig } from 'redux-persist';
-import { SliceConfigurator, buildSliceConfigurator } from '../../middleware/slice-configurator.js';
+import { SliceConfigurator, buildSliceConfigurator } from '../slice-configurator.js';
 import { EVENTS_INITIAL_STATE, eventsReducer, setupEventsListeners } from './events-slice.js';
 import { LIVE_GAME_INITIAL_STATE, LiveState, liveSlice } from './live-slice.js';
 import { shift } from './shift-slice.js';
